Remove AbstractComment
closes #1766
